@@ -101,7 +101,7 @@ def get_stats_for_timerange(
 @api_view(["GET"])
 def sentry_stats(request: HttpRequest):
     try:
-        current_time = datetime.utcnow()
+        current_time = datetime.now(datetime.UTC)
         target_end_date = current_time.strftime("%Y-%m-%dT%H:%M:%S")
         target_start_date = (current_time - timedelta(days=1)).strftime("%Y-%m-%dT%H:%M:%S")
 
