@@ -60,7 +60,7 @@ def check_condition(rollback_condition: dict, feature_flag: FeatureFlag) -> bool
         base_start_date = created_date.strftime("%Y-%m-%dT%H:%M:%S")
         base_end_date = (created_date + timedelta(days=1)).strftime("%Y-%m-%dT%H:%M:%S")
 
-        current_time = datetime.utcnow()
+        current_time = datetime.now(datetime.UTC)
         target_end_date = current_time.strftime("%Y-%m-%dT%H:%M:%S")
         target_start_date = (current_time - timedelta(days=1)).strftime("%Y-%m-%dT%H:%M:%S")
 
