@@ -65,8 +65,8 @@ class FormulaAST:
                 return operand
             raise ValueError(f"Operator {unary_op.__class__.__name__} not supported")
 
-        elif isinstance(node, ast.Num):
-            return node.n
+        elif isinstance(node, ast.Constant) and isinstance(node.value, int | float):
+            return node.value
 
         elif isinstance(node, ast.Name):
             try:
